added store_id in product table

+++ database/migrations/2022_05_09_123709_create_variants_table.php
@@ -24,9 +24,8 @@ class CreateVariantsTable extends Migration
             $table->bigInteger('product_id')->unsigned()->nullable();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
 
-            $table->bigInteger('variant_item_id')->unsigned()->nullable();
+            $table->Integer('variant_item_id')->unsigned()->nullable();
             $table->foreign('variant_item_id')->references('id')->on('variant_items')->onDelete('cascade');
-
 
             $table->string('quantity')->nullable();
             $table->string('price')->nullable();
